requirements.txt
asgiref==3.9.1
Django==3.2.16
django-filter==22.1
django-smart-selects==1.7.2
djangorestframework==3.14.0
pytz==2025.2
sqlparse==0.5.3